refactor: one export per file across the catalog sync
The script directory kept a shared `types.ts` and several modules exporting two
or three things, against the convention every file under `ai-models/types` and
`ai-models/utils` already follows. Each export now has its own file, split into
`types/` and `utils/` to match.

`ModelsDevCost` moves out of `models-dev-model.type.ts` for the same reason: I
added it there rather than beside it.

No behaviour change. The sync produces byte-identical artifacts and the 61
tests pass unchanged.
