--- packages/twenty-server/scripts/ai-catalog-sync/match-benchmarks.ts
@@ -1,153 +1,0 @@
-import { isDefined } from 'twenty-shared/utils';
-
-import { type AiModelBenchmarks } from 'src/engine/metadata-modules/ai/ai-models/types/ai-model-benchmarks.type';
-import { type ModelsDevModel } from 'src/engine/metadata-modules/ai/ai-models/types/models-dev-model.type';
-
-import { normalizeModelName } from './normalize-model-name.util';
-import {
-  type BenchmarkIndex,
-  type BenchmarkRecord,
-  type ObservedPrices,
-} from './types';
-
-const DATE_SUFFIX = /-\d{4}-?\d{2}-?\d{2}$/;
-const ROLLING_SUFFIX = /-(latest|preview|exp)$/;
-
-const fingerprint = (model: ModelsDevModel): string =>
-  JSON.stringify([model.cost ?? {}, model.limit ?? {}]);
-
-// `mistral-large-latest` is never benchmarked under that name, so it is
-// resolved to the dated release it currently points at by matching price and
-// limits against its non-rolling siblings. Re-resolves on its own when the
-// provider repoints the alias.
-const resolveRollingAlias = (
-  modelName: string,
-  siblingModels: Record<string, ModelsDevModel>,
-): string | undefined => {
-  const model = siblingModels[modelName];
-
-  if (!isDefined(model) || !ROLLING_SUFFIX.test(modelName)) {
-    return undefined;
-  }
-
-  const target = fingerprint(model);
-
-  const twins = Object.entries(siblingModels)
-    .filter(
-      ([siblingName, sibling]) =>
-        siblingName !== modelName &&
-        !ROLLING_SUFFIX.test(siblingName) &&
-        fingerprint(sibling) === target,
-    )
-    .sort(([, a], [, b]) =>
-      (b.release_date ?? '').localeCompare(a.release_date ?? ''),
-    );
-
-  return twins[0]?.[0];
-};
-
-export const buildLookupCandidates = (
-  modelName: string,
-  siblingModels: Record<string, ModelsDevModel>,
-): string[] => {
-  const candidates = [modelName, modelName.replace(DATE_SUFFIX, '')];
-
-  const resolved = resolveRollingAlias(modelName, siblingModels);
-
-  if (isDefined(resolved)) {
-    candidates.push(resolved, resolved.replace(DATE_SUFFIX, ''));
-  }
-
-  // Last: an undated `mistral-large` row in the index is whichever release the
-  // publisher last measured, so it must not win over the release the alias
-  // currently resolves to.
-  candidates.push(modelName.replace(ROLLING_SUFFIX, ''));
-
-  return [...new Set(candidates)];
-};
-
-const lookup = (
-  index: BenchmarkIndex,
-  candidates: string[],
-): BenchmarkRecord | undefined => {
-  for (const candidate of candidates) {
-    const record = index.get(normalizeModelName(candidate));
-
-    if (isDefined(record)) {
-      return record;
-    }
-  }
-
-  return undefined;
-};
-
-export type MatchBenchmarksArgs = {
-  modelName: string;
-  siblingModels: Record<string, ModelsDevModel>;
-  benchmarkIndex: BenchmarkIndex;
-  measuredAt: string;
-};
-
-export type BenchmarkMatch = {
-  // Absent when the publisher lists the model but has measured nothing about
-  // it, which is not the same as a model it scored badly.
-  benchmarks?: AiModelBenchmarks;
-  aliases: string[];
-  observedPrices?: ObservedPrices;
-};
-
-export const matchBenchmarks = ({
-  modelName,
-  siblingModels,
-  benchmarkIndex,
-  measuredAt,
-}: MatchBenchmarksArgs): BenchmarkMatch | undefined => {
-  const record = lookup(
-    benchmarkIndex,
-    buildLookupCandidates(modelName, siblingModels),
-  );
-
-  if (!isDefined(record)) {
-    return undefined;
-  }
-
-  const {
-    intelligenceIndex,
-    outputTokensPerSecond,
-    timeToFirstTokenSeconds,
-    costPerTask,
-    observedPrices,
-  } = record;
-
-  const hasMeasurement = [
-    intelligenceIndex,
-    outputTokensPerSecond,
-    timeToFirstTokenSeconds,
-    costPerTask,
-  ].some(isDefined);
-
-  // The publisher lists plenty of models it has measured nothing about, and a
-  // row of pure aliases says nothing worth carrying into either artifact.
-  const hasObservedPrice = [
-    observedPrices?.inputPerMillionTokens,
-    observedPrices?.outputPerMillionTokens,
-  ].some(isDefined);
-
-  if (!hasMeasurement && !hasObservedPrice) {
-    return undefined;
-  }
-
-  return {
-    benchmarks: hasMeasurement
-      ? {
-          intelligenceIndex,
-          outputTokensPerSecond,
-          timeToFirstTokenSeconds,
-          costPerTask,
-          measuredAt: record.measuredAt ?? measuredAt,
-        }
-      : undefined,
-    aliases: [...new Set([modelName, ...record.aliases])],
-    observedPrices: hasObservedPrice ? observedPrices : undefined,
-  };
-};
