--- packages/twenty-server/scripts/ai-catalog-sync/types.ts
@@ -1,55 +1,0 @@
-import { type AiSdkPackage } from 'twenty-shared/ai';
-
-import { type AiModelBenchmarks } from 'src/engine/metadata-modules/ai/ai-models/types/ai-model-benchmarks.type';
-import { type LongContextCost } from 'src/engine/metadata-modules/ai/ai-models/types/long-context-cost.type';
-import { type ModelFamily } from 'src/engine/metadata-modules/ai/ai-models/types/model-family.enum';
-
-export type GeneratedModel = {
-  name: string;
-  label: string;
-  description?: string;
-  modelFamily?: ModelFamily;
-  inputCostPerMillionTokens?: number;
-  outputCostPerMillionTokens?: number;
-  cachedInputCostPerMillionTokens?: number;
-  cacheCreationCostPerMillionTokens?: number;
-  longContextCost?: LongContextCost;
-  contextWindowTokens?: number;
-  maxOutputTokens?: number;
-  modalities?: string[];
-  supportsReasoning?: boolean;
-  benchmarks?: AiModelBenchmarks;
-  isDeprecated?: boolean;
-};
-
-export type GeneratedProvider = {
-  npm: AiSdkPackage;
-  label: string;
-  apiKey: string;
-  models: GeneratedModel[];
-};
-
-export type GeneratedCatalog = Record<string, GeneratedProvider>;
-
-export type ObservedPrices = {
-  inputPerMillionTokens?: number;
-  outputPerMillionTokens?: number;
-};
-
-export type BenchmarkRecord = {
-  intelligenceIndex?: number;
-  outputTokensPerSecond?: number;
-  timeToFirstTokenSeconds?: number;
-  costPerTask?: number;
-  // Kept out of the catalog on purpose: an observation only corroborates
-  // models.dev while it stays independent of it.
-  observedPrices?: ObservedPrices;
-  // Set only on a record recovered from the committed overlay, so preserved
-  // measurements keep the date they were actually taken.
-  measuredAt?: string;
-  aliases: string[];
-};
-
-// Keyed by normalized model name, so a single lookup resolves every alias a
-// source publishes for the same underlying model.
-export type BenchmarkIndex = Map<string, BenchmarkRecord>;
