+++ packages/twenty-server/scripts/ai-catalog-sync/types/benchmark-index.type.ts
@@ -1,0 +1,5 @@
+import { type BenchmarkRecord } from './benchmark-record.type';
+
+// Keyed by normalized model name, so a single lookup resolves every alias a
+// source publishes for the same underlying model.
+export type BenchmarkIndex = Map<string, BenchmarkRecord>;
