+++ packages/twenty-server/scripts/ai-catalog-sync/types/coverage-report.type.ts
@@ -1,0 +1,6 @@
+export type CoverageReport = {
+  generalPurposeModelCount: number;
+  scoredGeneralPurposeModelCount: number;
+  unscoredGeneralPurposeModelIds: string[];
+  specializedModelCount: number;
+};
