+++ packages/twenty-server/scripts/ai-catalog-sync/types/generated-catalog.type.ts
@@ -1,0 +1,3 @@
+import { type GeneratedProvider } from './generated-provider.type';
+
+export type GeneratedCatalog = Record<string, GeneratedProvider>;
