+++ packages/twenty-server/scripts/ai-catalog-sync/types/generated-model.type.ts
@@ -1,0 +1,21 @@
+import { type AiModelBenchmarks } from 'src/engine/metadata-modules/ai/ai-models/types/ai-model-benchmarks.type';
+import { type LongContextCost } from 'src/engine/metadata-modules/ai/ai-models/types/long-context-cost.type';
+import { type ModelFamily } from 'src/engine/metadata-modules/ai/ai-models/types/model-family.enum';
+
+export type GeneratedModel = {
+  name: string;
+  label: string;
+  description?: string;
+  modelFamily?: ModelFamily;
+  inputCostPerMillionTokens?: number;
+  outputCostPerMillionTokens?: number;
+  cachedInputCostPerMillionTokens?: number;
+  cacheCreationCostPerMillionTokens?: number;
+  longContextCost?: LongContextCost;
+  contextWindowTokens?: number;
+  maxOutputTokens?: number;
+  modalities?: string[];
+  supportsReasoning?: boolean;
+  benchmarks?: AiModelBenchmarks;
+  isDeprecated?: boolean;
+};
