+++ packages/twenty-server/scripts/ai-catalog-sync/types/generated-provider.type.ts
@@ -1,0 +1,10 @@
+import { type AiSdkPackage } from 'twenty-shared/ai';
+
+import { type GeneratedModel } from './generated-model.type';
+
+export type GeneratedProvider = {
+  npm: AiSdkPackage;
+  label: string;
+  apiKey: string;
+  models: GeneratedModel[];
+};
