+++ packages/twenty-server/scripts/ai-catalog-sync/types/match-benchmarks-args.type.ts
@@ -1,0 +1,10 @@
+import { type ModelsDevModel } from 'src/engine/metadata-modules/ai/ai-models/types/models-dev-model.type';
+
+import { type BenchmarkIndex } from './benchmark-index.type';
+
+export type MatchBenchmarksArgs = {
+  modelName: string;
+  siblingModels: Record<string, ModelsDevModel>;
+  benchmarkIndex: BenchmarkIndex;
+  measuredAt: string;
+};
