+++ packages/twenty-server/scripts/ai-catalog-sync/types/observed-prices.type.ts
@@ -1,0 +1,4 @@
+export type ObservedPrices = {
+  inputPerMillionTokens?: number;
+  outputPerMillionTokens?: number;
+};
