+++ packages/twenty-server/scripts/ai-catalog-sync/utils/build-catalog.util.ts
@@ -8,7 +8,8 @@ import { type ModelsDevData } from 'src/engine/metadata-modules/ai/ai-models/typ
 import { type ModelsDevModel } from 'src/engine/metadata-modules/ai/ai-models/types/models-dev-model.type';
 import { inferModelFamily } from 'src/engine/metadata-modules/ai/ai-models/utils/infer-model-family.util';
 
-import { type GeneratedCatalog, type GeneratedModel } from './types';
+import { type GeneratedCatalog } from '../types/generated-catalog.type';
+import { type GeneratedModel } from '../types/generated-model.type';
 
 const EXCLUDED_MODEL_PREFIXES = [
   'text-embedding',
@@ -136,21 +137,6 @@ const buildModel = ({
   }
 
   return model;
-};
-
-// A 200 carrying an empty or reshaped payload would otherwise generate an
-// empty catalog, and the sync PR automerges over the real one.
-export const assertPayloadIsUsable = (data: ModelsDevData): void => {
-  const missing = NATIVE_AI_SDK_PROVIDER_IDS.filter(
-    (providerName) =>
-      Object.keys(data[providerName]?.models ?? {}).length === 0,
-  );
-
-  if (missing.length > 0) {
-    throw new Error(
-      `models.dev payload is missing models for ${missing.join(', ')}; refusing to overwrite the catalog`,
-    );
-  }
 };
 
 export const buildCatalog = (data: ModelsDevData): GeneratedCatalog => {
