+++ packages/twenty-server/scripts/ai-catalog-sync/utils/match-benchmarks.util.ts
@@ -1,0 +1,80 @@
+import { isDefined } from 'twenty-shared/utils';
+
+import { type BenchmarkIndex } from '../types/benchmark-index.type';
+import { type BenchmarkMatch } from '../types/benchmark-match.type';
+import { type BenchmarkRecord } from '../types/benchmark-record.type';
+import { type MatchBenchmarksArgs } from '../types/match-benchmarks-args.type';
+
+import { buildLookupCandidates } from './build-lookup-candidates.util';
+import { normalizeModelName } from './normalize-model-name.util';
+
+const lookup = (
+  index: BenchmarkIndex,
+  candidates: string[],
+): BenchmarkRecord | undefined => {
+  for (const candidate of candidates) {
+    const record = index.get(normalizeModelName(candidate));
+
+    if (isDefined(record)) {
+      return record;
+    }
+  }
+
+  return undefined;
+};
+
+export const matchBenchmarks = ({
+  modelName,
+  siblingModels,
+  benchmarkIndex,
+  measuredAt,
+}: MatchBenchmarksArgs): BenchmarkMatch | undefined => {
+  const record = lookup(
+    benchmarkIndex,
+    buildLookupCandidates(modelName, siblingModels),
+  );
+
+  if (!isDefined(record)) {
+    return undefined;
+  }
+
+  const {
+    intelligenceIndex,
+    outputTokensPerSecond,
+    timeToFirstTokenSeconds,
+    costPerTask,
+    observedPrices,
+  } = record;
+
+  const hasMeasurement = [
+    intelligenceIndex,
+    outputTokensPerSecond,
+    timeToFirstTokenSeconds,
+    costPerTask,
+  ].some(isDefined);
+
+  // The publisher lists plenty of models it has measured nothing about, and a
+  // row of pure aliases says nothing worth carrying into either artifact.
+  const hasObservedPrice = [
+    observedPrices?.inputPerMillionTokens,
+    observedPrices?.outputPerMillionTokens,
+  ].some(isDefined);
+
+  if (!hasMeasurement && !hasObservedPrice) {
+    return undefined;
+  }
+
+  return {
+    benchmarks: hasMeasurement
+      ? {
+          intelligenceIndex,
+          outputTokensPerSecond,
+          timeToFirstTokenSeconds,
+          costPerTask,
+          measuredAt: record.measuredAt ?? measuredAt,
+        }
+      : undefined,
+    aliases: [...new Set([modelName, ...record.aliases])],
+    observedPrices: hasObservedPrice ? observedPrices : undefined,
+  };
+};
