+++ packages/twenty-server/scripts/ai-catalog-sync/utils/normalize-model-name.util.ts
