+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/models-dev-cost.type.ts
@@ -1,0 +1,6 @@
+export type ModelsDevCost = {
+  input?: number;
+  output?: number;
+  cache_read?: number;
+  cache_write?: number;
+};
